+++ fivetran/framework/core/model/destination_resource_model.go
@@ -18,6 +18,8 @@ type DestinationResourceModel struct {
 	DaylightSavingTimeEnabled	 types.Bool   	`tfsdk:"daylight_saving_time_enabled"`
 	Config  					 types.Object   `tfsdk:"config"`
 	Timeouts					 timeouts.Value `tfsdk:"timeouts"`
+    LocalProcessingAgentId       types.String   `tfsdk:"local_processing_agent_id"`
+    NetworkingMethod             types.String   `tfsdk:"networking_method"`
 
 	RunSetupTests    			 types.Bool 	`tfsdk:"run_setup_tests"`
 	TrustCertificates			 types.Bool 	`tfsdk:"trust_certificates"`
@@ -46,6 +48,18 @@ func (d *DestinationResourceModel) SetSetupStatus(value string) {
 }
 func (d *DestinationResourceModel) SetDaylightSavingTimeEnabled(value bool) {
 	d.DaylightSavingTimeEnabled = types.BoolValue(value)
+}
+func (d *DestinationResourceModel) SetLocalProcessingAgentId(value string) {
+    if value != "" {
+        d.LocalProcessingAgentId = types.StringValue(value)
+    } else {
+        d.LocalProcessingAgentId = types.StringNull()
+    }
+}
+func (d *DestinationResourceModel) SetNetworkingMethod(value string) {
+    if value != "" {
+        d.NetworkingMethod = types.StringValue(value)
+    }
 }
 func (d *DestinationResourceModel) SetConfig(value map[string]interface{}) {
 	if d.Service.IsNull() || d.Service.IsUnknown() {