+++ fivetran/framework/core/schema/connector.go
@@ -109,6 +109,10 @@ func ConnectorAttributesSchema() core.Schema {
 				ValueType:   	core.StringEnum,
 				Description: 	"Possible values: Directly, PrivateLink, SshTunnel, ProxyAgent.",
 			},
+			"local_processing_agent_id": {
+				ValueType:   core.String,
+				Description: "The local processing agent ID that refers to the controller created for the group the connection belongs to. If the value is specified, the system will try to associate the connection with an existing agent.",
+			},
 		},
 	}
 }